frontend/requirements.txt
streamlit
pandas
plotly
supabase-py
python-dotenv